// src/types.ts
export type templateType =
  | "next-js-auth0-tailwind"
  | "next-js-tailwind-css"
  | "next-js-tailwind-datocms"
  | "next-js-ts-tailwind"
  | "react-js"
  | "react-js-with-typescript"
  | "react-native-typescript"
  | "typescript-project"
  | "solid-js"
  | "solid-js-with-typescript";

export type pkgManagerType = "npm" | "yarn" | "pnpm";

export type choice = "Yes" | "No";